Refactor store creation into its own file
